Rename addr to bluer for clearer attribution
